Cast to int to fix failing Mac builds

+++ Moco/Moco/MocoGoal/MocoAccelerationTrackingGoal.cpp
@@ -125,7 +125,7 @@ void MocoAccelerationTrackingGoal::printDescriptionImpl(
     stream << "acceleration reference file: " 
            << get_acceleration_reference_file()
            << std::endl;
-    for (int i = 0; i < m_frame_paths.size(); i++) {
+    for (int i = 0; i < (int)m_frame_paths.size(); i++) {
         stream << "        ";
         stream << "frame " << i << ": " << m_frame_paths[i] << ", ";
         stream << "weight: " << m_acceleration_weights[i] << std::endl;
